+++ src/app/shopusa/comision/pages/comision-list/comision-list.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { Comision } from '../../../interfaces/comision.interface';
 import { ComisionService } from '../../services/comision.service';
 import { HttpResponse } from '@angular/common/http';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-comision-list',
@@ -16,15 +17,30 @@ export class ComisionListComponent implements OnInit {
   ngOnInit(): void {
     this.comisionService.getComisiones().subscribe((comisiones: Comision[]) => {
       this.comisiones = comisiones;
-      console.log(comisiones);
     });
   }
   deleteComision(id: string): void {
-    this.comisionService.deleteComision(id).subscribe((response) => {
-      console.log('Comision eliminada:', response);
-      this.comisiones = this.comisiones.filter(
-        (comision) => comision.id !== id
-      );
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: 'No podras deshacer esto.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.comisionService.deleteComision(id).subscribe((response) => {
+          this.comisiones = this.comisiones.filter(
+            (comision) => comision.id !== id,
+          );
+        });
+        Swal.fire({
+          title: 'Eliminado!',
+          text: 'Comision eliminada exitosamente.',
+          icon: 'success',
+        });
+      }
     });
   }
   generateXlsx(): void {
