+++ src/app/shopusa/plataformas/pages/plataforma-list/plataforma-list.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { Plataforma } from '../../../interfaces/plataforma.interface';
 import { PlataformaService } from '../../services/plataforma.service';
 import { HttpResponse } from '@angular/common/http';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-plataforma-list',
@@ -18,11 +19,28 @@ export class PlataformaListComponent implements OnInit {
     });
   }
   deletePlataforma(id: string): void {
-    this.plataformaService.deletePlataforma(id).subscribe((response) => {
-      console.log('Plataforma eliminada:', response);
-      this.plataformas = this.plataformas.filter(
-        (plataforma) => plataforma.id !== id
-      );
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: 'No podras deshacer esto.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.plataformaService.deletePlataforma(id).subscribe((response) => {
+          console.log('Plataforma eliminada:', response);
+          this.plataformas = this.plataformas.filter(
+            (plataforma) => plataforma.id !== id,
+          );
+        });
+        Swal.fire({
+          title: 'Eliminado!',
+          text: 'Plataforma eliminada exitosamente.',
+          icon: 'success',
+        });
+      }
     });
   }
   generateXlsx(): void {
